working

+++ experiments/fix_pcap.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
-in_pcap="$1"
+in_pcap=$1
 
-#echo $in_pcap
-#apt install tshark tcpreplay
-editcap -r $in_pcap input.pcap 1-5000  						# pega os 5000 primeiros pacotes do pcap recebido como parâmetro e joga em "input.pcap"
-tcprewrite --mtu=1500 --mtu-trunc --infile=input.pcap --outfile=output.pcap  	# trunca os pacotes do pacp recém criado para MTU=1500 bytes e joga em "output.pcap"
+if [ -z "$in_pcap" ]; then
+    echo No pcap provided
+    exit 1
+fi
+
+echo $in_pcap
+tshark -r $in_pcap -Y "tcp and (tcp.len > 224)" -w input.pcap
+editcap -r $in_pcap input.pcap 1-5000
+tcprewrite --mtu=1500 --mtu-trunc --infile=input.pcap --outfile=benign.pcap
 rm input.pcap
